Double default memory limit. Tame abort test.
With Linux 4.7 the value of RLIMIT_DATA is respected in calls to mmap
- a process wanting to allocate (in sum) more memory than RLIMIT_DATA
specifies will not be able to do so.

In FTPServer.pm the default limit is set to 16384 kilobytes.
This commit doubles the default memory limit so t/170files.t does not
fail anymore.

t/240abort.t originally constructs a huge string to simulate a big
file, also breaking the RLIMIT_DATA set.
This commit cuts down the string to a sane size not breaking the limit.